+++ src/Looper.hpp
@@ -263,12 +263,12 @@ struct Looper : Module {
 
       // Process each polyphony channel
 
-      for (size_t channel = 0; channel < loop.getChannels(p); channel++) {
+      for (int channel = 0; channel < loop.getChannels(p); channel++) {
         float in = inputs[ins[p]].getVoltage(channel);
         float rtrn = inputs[rtrns[p]].getVoltage(channel);
 
         float sample = loop.read(p, channel);
-        float rtrnGate = rtrnActive && inputs[rtrns[p]].getChannels() >= (channel + 1) ? mod : 0.0f;
+        float rtrnGate = rtrnActive && inputs[rtrns[p]].getChannels() >= (signed)(channel + 1) ? mod : 0.0f;
         float newSample = rtrnGate * rtrn + (1 - rtrnGate) * sample;
 
         loop.write(p, channel, feedback * newSample + inGate * in);
